refactor: Extract LruCache to top-level class, bound Color.lab_cache
Move LruCache from Palette inner class to ChromaWave::LruCache so it
can be shared. Replace unbounded Hash in Color.lab_cache with an
LRU-bounded cache to prevent growth from diverse #to_lab callers.
